Fixed fileMatch values in sarif-1.0.0.json
I also think the $schema property in this file is missused. 

I believe it is currently saying this file IS a schema and that it uses the custom schema format "http://json.schemastore.org/schema-catalog"
The docs in the standard are a little difficult to read, but this link describes it better https://spacetelescope.github.io/understanding-json-schema/reference/schema.html

Where as I belive your intent is to say this is a data file and it conforms to the JSON schemea found at "http://json.schemastore.org/schema-catalog". However as far as I'm aware their is no formal way to specifiy this.