Add support for ARM64 platform in AppVeyor schema
The ARM64 platform is a relatively new platform that builds binaries for aarch64, and Win32 is a legacy (but still the default) platform which builds 32-bits C++ apps. If you point Appveyor at a Visual C++ project file directly instead of a Visual Studio solution, you'll want to be able to set the platform to Win32.