Restore whitesourceScan stage definition
Partially reverts https://github.com/SchemaStore/schemastore/pull/1268
The config for a current version of the pipeline shall not include the restored parameters, however, older versions of the pipeline validate the config against the current version of the schema, which is an unfortunate oversight. So please merge as soon as possible, we have no other way to fix pipeline for clients. :-(