Fix added link for bitbucket pipelines manifest
Previously in 09da22f the bitbucket pipelines manifest JSON schema has
been linked in catalog.json for the first time.

It introduced two file-names matching the schema of which one was wrong.

The wrong file-name is with the .yaml extension, the correct name is with
the .yml extension.

Fix is to remove the wrong file-name from the catalogs' file-match array
for bitbucket-pipelines.yml JSON schema.

JSON-Catalog-Name: bitbucket-pipelines
Ref: 09da22fa5688395e01c72ecbe2633a005f2f237b
Caused-By: #1126