Add mandatory $schema present check (#1461)
* schema ID is now mandatory

Also implement "unknownFormat" for schemasafe

* license.1.json add "$schema"

* apibuilder.json add "$schema"

* apibuilder.json fix $ref issues

Keywords are ignore when place next tp $ref.
This is for standard 04,06 and 07
This behaviour is change in the newer standard.

* cosmos-config.json remove unsused additionalProperties

There is already a patternProperties.
It is not posible to add additionalProperties

* chrome-manifest.json remove minimum and maximum for enum[]

Enum is already a fix constant.

* chrome-manifest.json fix wrong object + items combination

There is a positive test for this. It pass the tv4 validator.
The tv4 does not see this error.

The schemasafe give error because of the object + items combination

* chrome-manifest.json add to "unknownFormat" list

Some unknown format is being used here.

* undo external schema commit.

These files are downloaded from external sources.
It should never updated/commit in the schemastore.