add: grunt task "local_count_schema_versions"
also add the "remote_count_schema_versions"
And at the build time show all the $schema version of all the files present
Schemasave and tv4 files are both scan for this.

There is only schema version info file for the schemasafe validator
and not for grunt-tv4 task.
Because grunt-tv4 have it own dedicated print output.