Add library definition to bukkit-plugin schema
In a recent commit to the bukkit api, an automatic library framework was
introduced that allows plugins to specify libraries hosted on maven
central which will be downloaded automatically on server start.
This framework can decrease the plugin jar size by avoiding shading
external libraries and instead shifting this responsibility to the
server itself.

These libraries are defined in the bukkit plugin.yml file, which follows
the bukkit-plugin.json schema.

This commit hence adds the schema specifications for the library
declarations to the bukkit-plugin.json file.
The individual library strings follow the pattern used by the maven
artifact resolver's DefaultArtifact.java class to validate artifacts.