github-workflow.json: fix "anyOf" to just "type": "object"
The previeus commit was change
from "type": "object" to "anyOf" due to broken build with AJV
It pass the build + test file. But still not correct.
It is now revert back to "type": "object" with the correct fix.

With negative test that add unknown keyword should have intercept this error.