Correct type schema for GitHub Actions workflow's /on/push
A "type" schema was added for the /on/push property of GitHub Actions workflow specifying that the type is "object", but
a "null" type is also allowed. This is already defined via the property's referenced "#/definitions/ref" schema, but the
newly added schema overrode that, causing spurious validation failures of valid workflows that used a null value for the
property.

I added a test for use of null values in all /on/* properties that support it (only the schedule event does not).