add: test module_copy.json
issue #1256
Ansible template type has an attribute validate which is set to boolean
in the JSON schema, but should be string.
The copy one has the same validate type and should be checked
and also corrected while working on this.

But there is no "validate" proprety in the module.copy schema.
I do not know how this is posible. This schema ia auto gererated from the
documentation.

Add test module_copy.json to check if any this else is wrong.
The test module_copy.json does not detect this missing property
because there is no "additionalProperties": false for this "properties"