Split GitHub Workflow schema job definitions
There are two flavors of jobs: "normal" jobs, and jobs which call
reusable workflows. Split the two cases using a oneOf schema.

In the previously existing case of a "normal" job, "runs-on" is
required and "uses" is forbidden. In the new case, the inverse is
true: "uses" is required and "runs-on" is forbidden.

The job types are moved into the definitions for the schema, and then
written under the desired `oneOf` block. The "needs" property, shared
between the two, is also moved to a new defintion for reuse.