prometheus.rules.json: allow multiple duration multipliers (#2011)
This change looses the pattern constraint for a duration string. It's valid for a Prometheus duration to not only consist of one duration multiplier (y,d,h,....), but multiple as long as they are in the order year, day, hour, minute, second, millisecond, e.g. `1h30m` or `2d12h`.

The pattern has been updated to consist of multiple optional capture groups for each duration multiplier instead of requiring only a single one. This pattern alone would match empty strings as well, therefore a second constraint has been added to check that the string length is at least 1.