github-workflow: fix runs-on specification
normalJob."runs-on" is documented here:

https://docs.github.com/en/actions/learn-github-actions/workflow-syntax-for-github-actions#jobsjob_idruns-on

If is effectively defined as a disjunction of:

* an enum of specific platform values like "macos-10.15"...
* a self-hosted specification

The latter is specified in the form of list of string values, with the
first element "self-hosted". The list of values that follows effectively
forms a constraint that defines the set of self-hosted runners on which
a job should run. e.g.

["self-hosted", "linux", "arm64", "gpu"]

"linux" and "arm64" are well-defined platform constraints, and "gpu" is
a custom label.

The current specification for "runs-on", specifically the "self-hosted"
variant, basically requires a string list value, with "self-hosted" as
the first element. But it incorrectly specifies that
"additionalProperites" should be of type string. This is not valid when
it comes to specifying the constraints of an array type, and should
instead be a constraint on "additionalItems".

Fix that, and also add a negative test to verify the change.

(A side note, but the "anyOf" specification for "runs-on" is redunant.
The first element in the "anyOf" constrains that the array starts with
"self-hosted" and can be followed by any string value (because of custom
labels). This subsumes all remaining "anyOf" elements. As such, we could
simplify the constraint for "runs-on" significantly, but that is left
for a later PR).