Fixes to allow ninjs-2.0 schema to pass strict validation checks.
For the $ref to work, we have to add an "id" to the geojson schema.

If there's a way to allow ajv to import the geojson schema without
adding the "id" value to the target schema, we would love to hear
how it can be done!

This commit also adds some positive and negative unit tests for ninjs.