Allow local path in reusable github-workflow call
In a github workflow job which calls a reusable workflow, the "uses"
value used to require a path to a repo "@" a version (a ref).
GitHub has added support for local paths beginning with "./", so
update the pattern used to validate this value appropriately.

Update the positive test case for calling reusable workflows to
include this usage, and add negative test with an invalid path.