Allow github-workflows exprs in `matrix.include`
`strategy.matrix.include` could contain an expression string, so
convert the `include/exclude` definition to allow for expression
strings via a oneOf.

(Requested downstream in python-jsonschema/check-jsonschema#272)

In support of this change, I compared both copies of the `matrix`
definition and found them to be identical. Therefore, they have been
refactored into a definition for matrix config.

A new test for `include: ${{ ... }}` has been added to exercise this
new branching within the schema.