Support undocumented name/remote in buf.gen.yaml (#3312)
* Support undocumented name/remote in buf.gen.yaml

Noticed after #3310 was merged that a few older buf.gen.yaml files had
these undocumented properties still around - they're supported, so I
don't think they should be "deprecated". They're exclusive of each
other, but one of them must be set.

(Otherwise, #3310 is working great on my end integrated with my $EDITOR
- thanks for the hard work on this project!)

* Remove periods from comments

* Remove "oneOf" trick with required

Maybe there's a way to express this with JSON Schema, but I'm not sure
how.

* Add `additionalProperties: false` to plugin schema

We've described all of the supported properties here.

* Remove wrapper "oneOf"

This is incorrect in this context.