Add Conventional Changelog Configuration file schema
This commit adds the JSON schema for the Conventional Changelog
Configuration file. It is used by tools like standard-version and its
newer fork commit-and-tag-version to generate change logs.
The configuration file that these tools recognize and that utilize this
schema, is usually named .versionrc, .versionrc.json or .versionrc.js.
The JSON schema is directly from the
conventional-changelog/conventional-changelog-config-spec repository.