Stricten pattern of micro settings.json (#3719)
In #3717 the pattern of the Micro editor config file was changed from
`*.settings.json` to `settings.json`. The name `settings.json` is too
generic and is used by multiple tools, including Visual Studio Code.

Based on the links in the referenced pull request, this changes the
pattern to the more specific `**/micro/settings.json`.