Require self-reference to $schema for abc-supply-plan
This commit introduces a requirement for JSON files conforming to the abc-supply-plan schema to include an explicit self-reference declaration. Following the example set by https://json.schemastore.org/schema-catalog.json, this change allows applications to infer the version of the abc-supply-plan schema, facilitating version conversion.

Additionally, this update includes the self-referential schema in the test cases and incorporates a negative test case to validate the failure of self-reference using the $schema keyword.

Testing:
- npm run grunt -- --SchemaName=abc-supply-plan-2.0.0.json
- npm run grunt