github-workflow: move matrix object constraints to oneof subschema
In a discussion with Jeremy Fiel on the JSON Schema Slack, he advised
that the constraints on the matrix field that relate to the object type
in the oneof should be moved to the object subschema in the oneof.

Not only is this more correct (a point on which I defer to Jeremy) it is
more human readable: the patternProperties, additionalProperties and
minProperties only apply to the matrix value in the case that its value
is an object.