Change second @typescript-eslint/consistent-type-definitions option to string enum
As opposed to the undefined object that was declared in the schema, the rule[0] declares that  the second argument is a string, as supported by the playground[1]:

```
This rule accepts the following options:

type Options = ['interface' | 'type'];

const defaultOptions: Options = ['interface'];

"interface" (default): enforce using interfaces for object type definitions.
"type": enforce using types for object type definitions.
```

[0] https://typescript-eslint.io/rules/consistent-type-definitions/#options
[1] https://typescript-eslint.io/play/#ts=5.5.2&fileType=.ts&code=C4TwDgpgBAglC8UDeAoKUCGAuKBnYATgJYB2A5igL4oqnAQEBmGAxtAELJqY77HlUUQA&eslintrc=N4KABGBEBOCuA2BTAzpAXGUEKQAIBcBPABxQGNoBLY-AWhXkoDt8B6MgeyeUuX0Ra0ipWgBNEAM2aV8lLqgwBtSImjQO0SABoowxJAC64MAF8QJoA&tsconfig=&tokens=false