fix: prettier config inconsistencies (#4571)
* chore: run prettier fix

No changes just run prettier fix across the code base to ensure
formatting abides by the standard.

* chore: sync prettier version in pre commit check

Update version of prettier version in .pre-commit-config.yaml with
that from package.json.

* fix: sync pre commit and npm prettier config

Sync the config of prettier in the pre commit hook with that of npm so
we don't get conflicts between the two.