add allowUnusedPatches to pnpm-workspace.json
I wasn't sure of the convention for marking a property as deprecated, so I left `allowNonAppliedPatches` as is.

I was tempted to include in the description the version number in which the property was added, but I didn't see that as a convention. I think it would be helpful though!