Update the sub-schema for poetry. (#4670)
* Update the sub-schema for poetry.

`"type": "string"` was missing from a few arrays.

`"additionalProperties": false,` was missing from a few objects.

Replaced all "oneOf" with "anyOf". For most cases in this schema, both are identical,
but "anyOf" is simpler. Only for "poetry-script-table" there is a minor difference
and "anyOf" is the correct option.

For the 'source' property, I replaced the "if", "then" and "else" with equivalent
logic that avoids these constructs.

These issues showed up while converting the poetry JSON schema to a
[TOML schema](https://github.com/udifuchs/toml-schema)
that I am experimenting with.

One feature of the JSON to TOML conversion tool is that it complains when
the JSON schema is different from what the tool expects.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>