Fix github-workflow schema to avoid ambiguous $ref (#5027)
For the `pull_request`, `pull_request_target`, and `push` event types,
the schema was declared to both define properties and use a `$ref`
which defined properties (`#/definitions/ref`).

This usage is ambiguous at best, since it's unclear if the properties are
meant to be merged -- which is the intent and AJV's behavior -- or
overridden. These definitions are updated here to avoid using a $ref at
all, and greatly enhance the clarity of their semantics by nesting
`oneOf` and `allOf` in a way which better reflects the criteria being
declared.

The newly added negative test passes with AJV using the schema in
`master` but fails when run through `python-jsonschema`'s toolchain.

Specifically, a failure is easy to produce with

    pipx run check-jsonschema \
        --schemafile schemas/json/github-workflow.json \
        src/negative_test/github-workflow/bad_pull_request_event_declaration.yaml

With the schema updated to remove the unclear `$ref` loading, both tools
agree that the negative test input is incorrect.