Replace \r\r with \n\n in clang-format-21.x.json
Most other schemas seem to use \n for newlines within descriptions, so
this adjusts the clang-format schema to match that pattern.

I've left this as a separate commit so it's easy to diff the changes
between clang-format-18.x.json and clang-format-21.x.json without every
property being changed because of this refactoring.