feat(pytest): update config to v9
The Pytest v9 release adds support for properly typed (TOML) configs,
instead of the string-only `ini_options` config. Configs can keep using
the legacy ini options, or migrate to the new config, though not both.

I have updated the Pytest TOML to reflect this, and I have taken the
opportunity to update all of the config options so that they match more
closely the upstream docs.

Signed-off-by: JP-Ellis <josh@jpellis.me>